+++ pkg/webui/components/wizard/form/index.js
@@ -26,16 +26,18 @@ import PrevButton from './prev-button'
 import NextButton from './next-button'
 
 const WizardForm = React.forwardRef((props, ref) => {
-  const { validationSchema, validationContext, onSubmit, children, initialValues, error } = props
+  const { validationSchema, validationContext, onSubmit, children, initialValues } = props
   const context = useWizardContext()
-  const { onNextStep, currentStep, steps, snapshot, onComplete, completeMessage } = context
+  const { onNextStep, currentStepId, steps, snapshot, onComplete, completeMessage, error } = context
 
   const formRef = React.useRef(null)
   const combinedRef = useCombinedRefs(ref, formRef)
 
   const stepsCount = steps.length
-  const isFirstStep = stepsCount > 0 ? currentStep === 1 : true
-  const isLastStep = stepsCount > 0 ? currentStep === stepsCount : true
+  const currentStepIndex = steps.findIndex(({ id }) => id === currentStepId)
+
+  const isFirstStep = stepsCount > 0 ? currentStepIndex === 0 : true
+  const isLastStep = stepsCount > 0 ? currentStepIndex === stepsCount - 1 : true
 
   const formInitialValues = React.useMemo(
     () => validationSchema.cast(merge({}, initialValues, snapshot), { context: validationContext }),
@@ -85,7 +87,6 @@ const WizardForm = React.forwardRef((props, ref) => {
 
 WizardForm.propTypes = {
   children: PropTypes.oneOfType([PropTypes.arrayOf(PropTypes.node), PropTypes.node]).isRequired,
-  error: PropTypes.error,
   initialValues: PropTypes.shape({}),
   onSubmit: PropTypes.func,
   validationContext: PropTypes.shape({}),
@@ -99,7 +100,6 @@ WizardForm.defaultProps = {
   onSubmit: undefined,
   validationContext: {},
   initialValues: {},
-  error: undefined,
 }
 
 export default WizardForm