+++ pkg/webui/components/wizard/form/prev-button.js
@@ -23,13 +23,15 @@ import Message from '@ttn-lw/lib/components/message'
 
 import PropTypes from '@ttn-lw/lib/prop-types'
 
+import style from './form.styl'
+
 const m = defineMessages({
   prev: 'Previous',
 })
 
 const WizardPrevButton = props => {
   const { isFirstStep, validationSchema, validationContext } = props
-  const { onPrevStep, currentStep, steps } = useWizardContext()
+  const { onPrevStep, currentStepId, steps } = useWizardContext()
   const { values } = useFormContext()
 
   const handlePrevStep = React.useCallback(() => {
@@ -44,14 +46,13 @@ const WizardPrevButton = props => {
     return null
   }
 
-  const { title: prevMessage } = steps.find(({ stepNumber }) => stepNumber === currentStep - 1) || {
-    title: m.next,
-  }
+  const currentStepIndex = steps.findIndex(({ id }) => id === currentStepId)
+  const { title: prevMessage } = steps[currentStepIndex - 1] || { title: m.prev }
 
   return (
-    <Button secondary onClick={handlePrevStep} type="button">
+    <Button className={style.button} secondary onClick={handlePrevStep} type="button">
       <Button.Icon icon="keyboard_arrow_left" type="left" />
-      <Message content={prevMessage} />
+      <Message className={style.messagePrev} content={prevMessage} />
     </Button>
   )
 }