+++ sdk/js/src/util/constants.js
@@ -14,6 +14,11 @@
 
 export const STACK_COMPONENTS = ['as', 'is', 'ns', 'js', 'gs', 'edtc', 'qrg', 'gcs']
 
+export const STACK_COMPONENTS_MAP = STACK_COMPONENTS.reduce((acc, curr) => {
+  acc[curr] = curr
+  return acc
+}, {})
+
 export const URI_PREFIX_STACK_COMPONENT_MAP = {
   as: 'as',
   ns: 'ns',